move client methods to client.go, utils to new utils.go
